:sparkles: 调整首页快速切换
